// src/app/hotel/types.ts
export interface HotelDetailInstallmentData {
  headers: number[]
  items: {
    bank: string
    description: string
    logo: string
    installments: {
      [key: string]: number
    }
  }[]
}